+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## 0.0.6 (2020-xx-xx Unreleased)
 
-- coming soon...
+- run selected build
 
 ## 0.0.5 (2020-01-06)